# server-spring-boot/api-specifications.yaml
openapi: 3.0.1
info:
  title: Hostel Management System API (MongoDB)
  description: RESTful API for managing hostel operations including student registration, room allocation, food menu management, and room change requests. This version uses MongoDB as the database.
  contact:
    name: Hostel Management Team
    url: https://hostel.com
    email: admin@hostel.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Development Server (MongoDB)
- url: https://api.hostel.com
  description: Production Server (MongoDB)
security:
- Bearer Authentication: []
paths:
...
  /api/warden/create-student:
    post:
      tags:
      - Warden Management
      summary: Create New Student
      description: Create a new student account with personal details and room assignment
      operationId: createStudent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
        required: true
      responses:
        201:
          description: Student created successfully
          content:
            application/json:
              schema:
                example: {"message": "Student created successfully", "studentId": "123"}
        400:
          description: Invalid request data
          content:
            application/json:
              schema:
                example: {"error": "Username already exists"}
        401:
          description: Unauthorized - Warden access required
          content:
            application/json:
              schema:
                example: {"error": "Unauthorized"}
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                example: {"error": "Database error"}
      security:
      - Bearer Authentication: []
...
components:
  schemas:
...
    CreateStudentRequest:
      required:
      - email
      - fullName
      - password
      - phone
      - username
      type: object
      properties:
        username:
          type: string
          description: Student's username
        password:
          type: string
          description: Student's password
        fullName:
          type: string
          description: Student's full name
        email:
          type: string
          description: Student's email address
        phone:
          type: string
          description: Student's phone number
        roomNumber:
          type: string
          description: Room assignment
        bedNumber:
          type: integer
          description: Bed assignment
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter JWT Bearer token in the format - Bearer {token}
      scheme: bearer
      bearerFormat: JWT 